+++ specsim/camera.py
@@ -320,12 +320,13 @@ class Camera(object):
         """Downsample data tabulated on the simulation grid to output pixels.
         """
         data = np.asanyarray(data)
-        if data.shape != self._wavelength.shape:
+        if data.shape[-1] != len(self._wavelength):
             raise ValueError(
                 'Invalid data shape for downsampling: {0}.'.format(data.shape))
 
-        return method(
-            data[self._output_slice].reshape(self._downsampled_shape), axis=-1)
+        output = data[..., self._output_slice]
+        new_shape = output.shape[:-1] + self._downsampled_shape
+        return method(output.reshape(new_shape), axis=-1)
 
 
     def apply_resolution(self, flux):
@@ -337,8 +338,8 @@ class Camera(object):
         flux = np.asarray(flux)
         dispersed = np.zeros_like(flux)
 
-        dispersed[self.ccd_slice] = self._resolution_matrix.dot(
-            flux[self.response_slice])
+        dispersed[..., self.ccd_slice] = self._resolution_matrix.dot(
+            flux[..., self.response_slice].T).T
 
         return dispersed