+++ specsim/fiberloss.py
@@ -9,6 +9,8 @@ from __future__ import print_function, division
 import numpy as np
 import numpy.lib.stride_tricks
 
+import scipy.interpolate
+
 import astropy.units as u
 import astropy.table
 
@@ -276,6 +278,8 @@ class GalsimFiberlossCalculator(object):
 
 def calculate_fiber_acceptance_fraction(
     focal_x, focal_y, wavelength, source, atmosphere, instrument,
+    source_types=None, source_fraction=None, source_half_light_radius=None,
+    source_minor_major_axis_ratio=None, source_position_angle=None,
     oversampling=32, saved_images_file=None, saved_table_file=None):
     """Calculate the acceptance fraction for a single fiber.
 
@@ -299,19 +303,43 @@ def calculate_fiber_acceptance_fraction(
         X coordinate of the fiber center in the focal plane with length units.
     focal_y : :class:`astropy.units.Quantity`
         Y coordinate of the fiber center in the focal plane with length units.
-    wavelength : :class:`astropy.table.Column`
-        Array of simulation wavelengths where the fiber acceptance fraction
-        should be tabulated.
+    wavelength : :class:`astropy.units.Quantity`
+        Array of simulation wavelengths (with length units) where the fiber
+        acceptance fraction should be tabulated.
     source : :class:`specsim.source.Source`
         Source model to use for the calculation.
     atmosphere : :class:`specsim.atmosphere.Atmosphere`
         Atmosphere model to use for the calculation.
     instrument : :class:`specsim.instrument.Instrument`
         Instrument model to use for the calculation.
+    source_types : array or None
+        Array of source type names that identify which tabulated fiberloss
+        fraction should be used for each fiber with the ``table`` method.
+        Each name should already be defined as a key in the
+        ``instruments.fiberloss.table.paths`` configuration.
+        Ignored for the ``galsim`` method.
+    source_fraction : array or None
+        Array of shape (num_fibers, 2).  See
+        :meth:`GalsimFiberlossCalculator.create_source` for details.
+        Ignored for the ``table`` method.
+    source_half_light_radius : array or None
+        Array of shape (num_fibers, 2).  See
+        :meth:`GalsimFiberlossCalculator.create_source` for details.
+        Ignored for the ``table`` method.
+    source_minor_major_axis_ratio : array or None
+        Array of shape (num_fibers, 2).  See
+        :meth:`GalsimFiberlossCalculator.create_source` for details.
+        Ignored for the ``table`` method.
+    source_position_angle : array or None
+        Array of shape (num_fibers, 2).  See
+        :meth:`GalsimFiberlossCalculator.create_source` for details.
+        Ignored for the ``table`` method.
     oversampling : int
         Oversampling factor to use for anti-aliasing the fiber aperture.
+        Ignored for the ``table`` method.
     saved_images_file : str or None
         See :meth:`GalsimFiberlossCalculator.calculate`.
+        Ignored for the ``table`` method.
     saved_table_file : str or None
         Write a table of calculated values to a file with this name.  The
         extension determines the file format, and .ecsv is recommended.
@@ -324,15 +352,30 @@ def calculate_fiber_acceptance_fraction(
         Array of fiber acceptance fractions (dimensionless) at each of the
         input wavelengths.
     """
-    # Use pre-tabulated fiberloss vs wavelength when available.
-    num_wlen = instrument.fiberloss_num_wlen
-    if num_wlen == 0:
-        return instrument.fiber_acceptance_dict[source.type_name]
+    num_fibers = len(focal_x)
+    if len(focal_y) != num_fibers:
+        raise ValueError('Arrays focal_x and focal_y must have same length.')
+
+    # Use pre-tabulated fiberloss fractions when requested.
+    if instrument.fiberloss_method == 'table':
+        if source_types is None:
+            # Use same source type for all fibers.
+            return instrument.fiber_acceptance_dict[
+                source.type_name][np.newaxis, :]
+        elif len(source_types) != num_fibers:
+            raise ValueError('Unexpected shape for source_types.')
+        floss = np.empty((num_fibers, len(wavelength)))
+        for i, type_name in enumerate(source_types):
+            floss[i] = instrument.fiber_acceptance_dict[type_name]
+        return floss
+
+    # Otherwise, use GalSim to calculate fiberloss fractions on the fly...
 
     # Initialize the grid of wavelengths where the fiberloss will be
     # calculated.
+    num_wlen = instrument.fiberloss_num_wlen
     wlen_unit = wavelength.unit
-    wlen_grid = np.linspace(wavelength.data[0], wavelength.data[-1],
+    wlen_grid = np.linspace(wavelength.value[0], wavelength.value[-1],
                             num_wlen) * wlen_unit
 
     # Initialize a new calculator.
@@ -341,28 +384,43 @@ def calculate_fiber_acceptance_fraction(
         wlen_grid.to(u.Angstrom).value,
         instrument.fiberloss_num_pixels,
         oversampling,
-        atmosphere.seeing['moffat_beta'])
+        atmosphere.seeing_moffat_beta)
 
     # Calculate the focal-plane optics at the fiber locations.
     scale, blur, offset = instrument.get_focal_plane_optics(
-        focal_x.reshape(1,), focal_y.reshape(1,), wlen_grid)
+        focal_x, focal_y, wlen_grid)
 
     # Calculate the atmospheric seeing at each wavelength.
     seeing_fwhm = atmosphere.get_seeing_fwhm(wlen_grid).to(u.arcsec).value
 
-    # Lookup the source model parameters, which we assume to be achromatic.
-    source_fraction = np.empty((1, 2))
-    source_hlr = np.empty((1, 2))
-    source_q = np.empty((1, 2))
-    source_beta = np.empty((1, 2))
-    source_fraction[0, 0] = source.disk_fraction
-    source_fraction[0, 1] = source.bulge_fraction
-    source_hlr[0, 0] = source.disk_shape.half_light_radius.to(u.arcsec).value
-    source_hlr[0, 1] = source.bulge_shape.half_light_radius.to(u.arcsec).value
-    source_q[0, 0] = source.disk_shape.minor_major_axis_ratio
-    source_q[0, 1] = source.bulge_shape.minor_major_axis_ratio
-    source_beta[0, 0] = source.disk_shape.position_angle.to(u.deg).value
-    source_beta[0, 1] = source.bulge_shape.position_angle.to(u.deg).value
+    # Replicate source parameters from the source config if they are not
+    # provided via args. If they are provided, check for the expected shape.
+    if source_fraction is None:
+        source_fraction = np.tile(
+            [source.disk_fraction, source.bulge_fraction], [num_fibers, 1])
+    elif source_fraction.shape != (num_fibers, 2):
+        raise ValueError('Unexpected shape for source_fraction.')
+    if source_half_light_radius is None:
+        source_half_light_radius = np.tile(
+            [source.disk_shape.half_light_radius.to(u.arcsec).value,
+             source.bulge_shape.half_light_radius.to(u.arcsec).value],
+            [num_fibers, 1])
+    elif source_half_light_radius.shape != (num_fibers, 2):
+        raise ValueError('Unexpected shape for source_half_light_radius.')
+    if source_minor_major_axis_ratio is None:
+        source_minor_major_axis_ratio = np.tile(
+            [source.disk_shape.minor_major_axis_ratio,
+             source.bulge_shape.minor_major_axis_ratio],
+            [num_fibers, 1])
+    elif source_minor_major_axis_ratio.shape != (num_fibers, 2):
+        raise ValueError('Unexpected shape for source_minor_major_axis_ratio.')
+    if source_position_angle is None:
+        source_position_angle = np.tile(
+            [source.disk_shape.position_angle.to(u.deg).value,
+             source.bulge_shape.position_angle.to(u.deg).value],
+            [num_fibers, 1])
+    elif source_position_angle.shape != (num_fibers, 2):
+        raise ValueError('Unexpected shape for source_position_angle.')
 
     # Calculate fiberloss fractions.  Note that the calculator expects arrays
     # with implicit units.
@@ -370,10 +428,12 @@ def calculate_fiber_acceptance_fraction(
         seeing_fwhm,
         scale.to(u.um / u.arcsec).value, offset.to(u.um).value,
         blur.to(u.um).value,
-        source_fraction, source_hlr, source_q, source_beta,
+        source_fraction, source_half_light_radius,
+        source_minor_major_axis_ratio, source_position_angle,
         saved_images_file)
 
-    if saved_table_file:
+    # TODO: add support for saving table when num_fibers > 1.
+    if saved_table_file and num_fibers == 1:
         meta = dict(
             description='Fiberloss fraction for source "{0}"'
             .format(source.name) +
@@ -392,4 +452,11 @@ def calculate_fiber_acceptance_fraction(
         table.write(saved_table_file, **args)
 
     # Interpolate (linearly) to the simulation wavelength grid.
-    return np.interp(wavelength.data, wlen_grid.value, fiberloss_grid[0])
+    # Use scipy.interpolate instead of np.interp here to avoid looping
+    # over fibers.
+    interpolator = scipy.interpolate.interp1d(
+        wlen_grid.value, fiberloss_grid, kind='linear', axis=1,
+        copy=False, assume_sorted=True)
+    # Both wavelength grids have the same units, by construction, so no
+    # conversion factor is required here.
+    return interpolator(wavelength.value)